+++ appServices/music/app/services/genreSongService.py
@@ -1,13 +1,15 @@
 from app.repositories.genreSongRepo import GenreSongRepo
 from app.entities.genreSong import GenreSong
 from common.errorCodes import ErrorCodes
+from app.repositories.genresRepo import GenresRepo
+from app.repositories.songsRepo import SongsRepo
 import uuid
 
 class GenreSongService:
     @staticmethod
     def findAllPaginated(page: int = 1, pageSize: int = 10):
         try:
-            result = GenreSongRepo.getAllPaginated(page, pageSize)
+            result = GenreSongRepo.filterAllPaginated(page, pageSize)
             if not result:
                 return None, ErrorCodes.NOT_FOUND
             return result, None
@@ -17,7 +19,7 @@ class GenreSongService:
     @staticmethod
     def findAll():
         try:
-            result = GenreSongRepo.getAll()
+            result = GenreSongRepo.filterAll()
             if not result:
                 return None, ErrorCodes.NOT_FOUND
             return result, None
@@ -79,6 +81,14 @@ class GenreSongService:
         try:
             genreId = uuid.UUID(genreId)
             songId = uuid.UUID(songId)
+            genre = GenresRepo.getById(genreId)
+            if not genre:
+                return None, ErrorCodes.NOT_FOUND
+            song = SongsRepo.getById(songId)
+            if not song:
+                return None, ErrorCodes.NOT_FOUND
+            if GenreSongRepo.getExactly(genreId, songId):
+                return None, ErrorCodes.ALREADY_EXISTS
             genreSong = GenreSong(genreId=genreId, songId=songId)
             result = GenreSongRepo.create(genreSong)
             if not result:
@@ -93,6 +103,8 @@ class GenreSongService:
             genreId = uuid.UUID(genreId)
             songId = uuid.UUID(songId)
             genreSong = GenreSongRepo.getExactly(genreId, songId)
+            if not genreSong:
+                return None, ErrorCodes.NOT_FOUND
             result = GenreSongRepo.delete(genreSong)
             if not result:
                 return None, ErrorCodes.DELETE_FAILED
